requirements.txt
# --- Data Collection ---
yfinance>=0.2.52                  # Yahoo Finance OHLCV + SPY + VIX
pytrends>=4.9.2                   # Google Trends search interest
praw>=7.8.1                       # Reddit API wrapper (Python 3.9+)
fredapi>=0.5.2                    # FRED macroeconomic data

# --- Feature Engineering ---
ta==0.11.0                        # Technical indicators (RSI, MACD, Bollinger)
vaderSentiment==3.3.2             # VADER sentiment (latest stable, unchanged since 2018)

# --- Data Handling ---
pandas>=2.2.0,<3.0                # Pin below 3.0 — breaking changes in 3.0
numpy>=1.26.0,<2.0                # numpy 2.0 has breaking changes; pin 1.x
scipy>=1.13.0

# --- Machine Learning ---
scikit-learn>=1.4.0,<1.7          # kNN, NB, DT, RF, MLP + Pipeline + TimeSeriesSplit
imbalanced-learn>=0.12.0          # SMOTE (requires scikit-learn >=1.4)

# --- Explainability ---
shap>=0.46.0                      # TreeExplainer for Random Forest; 3.12 support from 0.44+

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0

# --- Agent / Dashboard ---
streamlit>=1.32.0                 # Streamlit live dashboard (Python 3.9+)

# --- Utilities ---
python-dotenv>=1.0.1              # Load local .env for API credentials
joblib>=1.4.2                     # Model bundle serialization